Move autofix-printing.ts from language/ to contracts/ directory
RATIONALE:
 was in , but its purpose is
not language-layer internals (grammar, recovery, etc.). It is a *public
contract* for rendering AST nodes back to GML source text for lint
autofixes — the same role as , which already lives
in .

The directory  holds cross-cutting public APIs
consumed by rules. Placing the printing contract there makes it:
1. Discoverable as a first-class artifact rather than hidden under
   language/ alongside unrelated recovery helpers.
2. Directly importable by rule authors without reaching through
   language/ internal structure.
3. Clearly demarcated from formatting logic (which belongs in
   @gmloop/format) and transpiler output logic.

WHAT CHANGED:
- src/lint/src/contracts/autofix-printing.ts — canonical implementation
  (new file)
- src/lint/src/contracts/index.ts — exports the three public functions
- src/lint/src/language/autofix-printing.ts — converted to backward-
  compatibility re-export shim (imports from contracts/)
- src/lint/src/rules/gml/gml-rule-services.ts — updated import path
  and JSDoc to reference the new canonical location

All internal call sites (gml-rule-services, language/index.ts barrel)
continue to work via the shim. Tests for the printing functions pass.
