+++ src/lint/src/contracts/index.ts
@@ -1,1 +1,2 @@
+export { printExpression, printNodeForAutofix, readNodeText } from "./autofix-printing.js";
 export type { LintPluginShape } from "./lint-plugin-shape.js";
